+++ apps/docs/src/app/core/component-docs/select/examples/select-custom-trigger/select-custom-trigger.component.ts
@@ -1,9 +1,10 @@
-import { Component } from '@angular/core';
+import { Component, ViewEncapsulation } from '@angular/core';
 
 @Component({
     selector: 'fd-select-custom-trigger',
     templateUrl: './select-custom-trigger.component.html',
-    styleUrls: ['./select-custom-trigger.component.scss']
+    styleUrls: ['./select-custom-trigger.component.scss'],
+    encapsulation: ViewEncapsulation.None
 })
 export class SelectCustomTriggerComponent {
     options: string[] = ['Apple', 'Pineapple', 'Tomato', 'Strawberry'];