+++ e2e/wdio/platform/pages/base-component.po.ts
@@ -1,4 +1,12 @@
-import { click, elementArray, scrollIntoView, open } from '../../driver/wdio';
+import {
+    click,
+    elementArray,
+    scrollIntoView,
+    open,
+    saveElementScreenshot,
+    getImageTagBrowserPlatform,
+    checkElementScreenshot, waitForElDisplayed
+} from '../../driver/wdio';
 import { checkLtrOrientation, checkRtlOrientation } from '../../helper/assertion-helper';
 
 export class BaseComponentPo {
@@ -17,6 +25,26 @@ export class BaseComponentPo {
             click(switchers, i);
             checkLtrOrientation(areas, i);
         }
+    }
+
+    saveExampleBaselineScreenshot(specName: string, options: object = {}, areas: string = this.exampleAreaContainersArr): void {
+        const areasArray = elementArray(areas);
+        for (let i = 0; i < areasArray.length; i++) {
+            waitForElDisplayed(areas, i);
+            scrollIntoView(areas, i);
+            saveElementScreenshot(areas, `${specName}-example-${i}-platform-${getImageTagBrowserPlatform()}`, options, i);
+        }
+    }
+
+    compareWithBaseline(specName: string, options: object = {}, areas: string = this.exampleAreaContainersArr): any {
+        const areasArray = elementArray(areas);
+        let diff = 0;
+        for (let i = 0; i < areasArray.length; i++) {
+            waitForElDisplayed(areas, i);
+            scrollIntoView(areas, i);
+            diff +=  checkElementScreenshot(areas, `${specName}-example-${i}-platform-${getImageTagBrowserPlatform()}`, options, i);
+        }
+        return diff;
     }
 
     open(url: string): void {