+++ e2e/wdio/platform/tests/menu.e2e-spec.ts
@@ -170,6 +170,13 @@ describe('Menu component test suite', function() {
             expect(getAttributeByName(menuPage.exampleAreaContainersArr, 'dir', i)).toBe('rtl');
         }
     });
+
+    describe('Check visual regression', function() {
+        it('should check examples visual regression', () => {
+            menuPage.saveExampleBaselineScreenshot('menu');
+            expect(menuPage.compareWithBaseline('menu')).toBeLessThan(1);
+        });
+    });
 });
 
 function checkMenuItemsHoverState(itemsArrSelector, attribute, expectation): void {