+++ e2e/wdio/platform/tests/standard-list-item.e2e-spec.ts
@@ -179,4 +179,11 @@ describe('Standard List Item test suite:', function() {
             standardListPg.checkRtlSwitch();
         });
     });
+
+    xdescribe('Check visual regression', function() {
+        it('should check examples visual regression', () => {
+            standardListPg.saveExampleBaselineScreenshot('standard-list-item', {removeElements: $(standardListPg.sFooterByLineAvatar)});
+            expect(standardListPg.compareWithBaseline('standard-list-item', {removeElements: $(standardListPg.sFooterByLineAvatar)})).toEqual(0);
+        });
+    });
 });