+++ libs/core/src/lib/switch/switch.component.ts
@@ -81,6 +81,14 @@ export class SwitchComponent implements ControlValueAccessor {
     @Input()
     ariaLabelledby: string = null;
 
+    /** Semantic Label Accept set for Accessibility */
+    @Input()
+    semanticAcceptLabel = 'Accept';
+
+    /** Semantic Label Decline set for Accessibility */
+    @Input()
+    semanticDeclineLabel = 'Decline';
+
     /**
      * Event fired when the state of the switch changes.
      * *$event* can be used to retrieve the new state of the switch.
@@ -89,21 +97,26 @@ export class SwitchComponent implements ControlValueAccessor {
     readonly checkedChange: EventEmitter<boolean> = new EventEmitter<boolean>();
 
     /** @hidden */
-    onChange: any = () => {};
+    onChange: Function = () => {};
 
     /** @hidden */
-    onTouched: any = () => {};
+    onTouched: Function = () => {};
 
-    constructor(private changeDetectorRef: ChangeDetectorRef) {}
+    constructor(private readonly _changeDetectorRef: ChangeDetectorRef) {}
 
     /** Set focus on the input element. */
-    public focus(): void {
+    focus(): void {
         this.inputElement.nativeElement.focus();
     }
 
     /** Get the id of the inner input element of the switch. */
     get innerInputId(): string {
         return `${this.id}-input`;
+    }
+
+    /** Get the id of the semantic label element of the switch. */
+    get _semanticLabelId(): string {
+        return `${this.id}-semantic-label`;
     }
 
     /** Get the isChecked property of the switch. */
@@ -125,7 +138,7 @@ export class SwitchComponent implements ControlValueAccessor {
      */
     writeValue(value: any): void {
         this.checked = value;
-        this.changeDetectorRef.detectChanges();
+        this._changeDetectorRef.detectChanges();
     }
 
     /**
@@ -140,7 +153,7 @@ export class SwitchComponent implements ControlValueAccessor {
      * @hidden
      * @param fn User defined function that handles the *onTouch* event of the switch.
      */
-    registerOnTouched(fn): void {
+    registerOnTouched(fn: Function): void {
         this.onTouched = fn;
     }
 
@@ -150,6 +163,6 @@ export class SwitchComponent implements ControlValueAccessor {
      */
     setDisabledState(isDisabled: boolean): void {
         this.disabled = isDisabled;
-        this.changeDetectorRef.detectChanges();
+        this._changeDetectorRef.detectChanges();
     }
 }